Hacky and likely inefficient fix for half precision

+++ rome/compute_u.py
@@ -139,7 +139,7 @@ def compute_u(
             hparams.mom2_dataset,
             hparams.mom2_n_samples,
             hparams.mom2_dtype,
-        ) @ u.unsqueeze(1)
+        ) @ u.float().unsqueeze(1)
         u = u.squeeze()
 
     return u / u.norm()
